color-management: Add HDR-metadata support
This patch adds HDR-metadata support in the color-management-protocol.
It enables a client to:
- know if an output is HDR capable.
- set HDR-metadata values, received from an HDR content, for a surface.

The HDR-metadata values : MAX_CLL, MAX_FALL, MAX_LUMINANCE,
MIN LUMINANCE can be sent from the client to compositor, which can
inturn send these to kernel. Kernel can finally send these values in
AVI-INFOFRAMES to the HDR display to provide an idea of the brightness
of the content. The display can use this information to adapt itself
for a better viewing experience.

Signed-off-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>